Add desired production and test code
This is a rip-off of the example in the googletest primer.

Test code should contain very little boilerplate. Ideally, it should be possible to put tests directly into packages (or modules). SVUnit should figure out the qualified test names. This is hopefully possible, as SystemVerilog provides some introspection capabilities (as opposed to C++).